+++ src/main/java/org/mo/jims/coop/service/SellService.java
@@ -9,15 +9,13 @@ import org.mo.jims.coop.entity.Sell;
 import org.mo.jims.coop.repository.SellRepository;
 import org.mo.open.common.util.Page;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Propagation;
-import org.springframework.transaction.annotation.Transactional;
 
 @Service("sellService")
 public class SellService {
 
 	private SellRepository sellRepository;
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public Page<Sell> getSellByCriteria(String goodName, String customerName,
 			Date beginTime, Date endTime, int page, int pageSize) {
 		Page<Sell> SellPage = new Page<Sell>();
@@ -35,12 +33,15 @@ public class SellService {
 		return SellPage;
 	}
 
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public Sell getSellByPK(String id) {
-		return sellRepository.selectByPK(id);
+		if(id!=null && !"".equals(id)){
+			return sellRepository.selectByPK(id);
+		}
+		return null;
 	}
 
-	@Transactional(rollbackFor = RuntimeException.class)
+	
 	public boolean saveSell(Sell entity) {
 		if (entity != null) {
 			if (entity.getTime() == null) {
@@ -52,109 +53,115 @@ public class SellService {
 		return false;
 	}
 
-	@Transactional(rollbackFor = RuntimeException.class)
+	
 	public boolean alterSell(Sell entity) {
-		sellRepository.updateByPK(entity);
-		return true;
+		if(entity!=null){
+			sellRepository.updateByPK(entity);
+			return true;
+		}
+		return false;
 	}
 
-	@Transactional(rollbackFor = RuntimeException.class)
+	
 	public boolean removeSellByPK(String id) {
-		sellRepository.deleteByPK(id);
-		return true;
+		if(id!=null && !"".equals(id)){
+			sellRepository.deleteByPK(id);
+			return true;
+		}
+		return false;
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public List<Sell> getSellInToday() {
 		return sellRepository.selectToday();
 	}
 
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public List<Sell> getSellInYesterday() {
 		return sellRepository.selectYesterday();
 	}
 
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public List<Sell> getSellInThisWeek() {
 		return sellRepository.selectThisWeek();
 	}
 
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public List<Sell> getSellInLastWeek() {
 		return sellRepository.selectLastWeek();
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public List<Sell> getSellInThisMonth() {
 		return sellRepository.selectThisMonth();
 	}
 
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public List<Sell> getSellInLastMonth() {
 		return sellRepository.selectLastMonth();
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public List<Sell> getSellInThisYear() {
 		return sellRepository.selectThisYear();
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public List<Sell> getSellInLastYear() {
 		return sellRepository.selectLastYear();
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public List<Sell> getAllSell(){
 		return sellRepository.selectAll();
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public int countSellInToday(){
 		return sellRepository.countToday();
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public int countSellInYesterday(){
 		return sellRepository.countYesterday();
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public int countSellInThisWeek() {
 		return sellRepository.countThisWeek();
 	}
 
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public int countSellInLastWeek() {
 		return sellRepository.countLastWeek();
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public int countSellInThisMonth() {
 		return sellRepository.countThisMonth();
 	}
 
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public int countSellInLastMonth() {
 		return sellRepository.countLastMonth();
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public int countSellInThisYear() {
 		return sellRepository.countThisYear();
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public int countSellInLastYear() {
 		return sellRepository.countLastYear();
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public int countAllSell(){
 		return sellRepository.countAll();
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public Float getSellPercentageofDayOnThisMonth(){
 		int countToday = sellRepository.countToday();
 		int countYesterday = sellRepository.countYesterday();
@@ -163,7 +170,7 @@ public class SellService {
 		return result;
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public Float getSellPercentageofWeekOnThisMonth(){
 		int countToday = sellRepository.countThisWeek();
 		int countYesterday = sellRepository.countLastWeek();
@@ -172,7 +179,7 @@ public class SellService {
 		return result;
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public Float getSellPercentageofMonthOnThisYear(){
 		int countToday = sellRepository.countThisMonth();
 		int countYesterday = sellRepository.countLastMonth();
@@ -181,7 +188,7 @@ public class SellService {
 		return result;
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public Float getSellPercentageofYearOnAll(){
 		int countToday = sellRepository.countThisYear();
 		int countYesterday = sellRepository.countLastYear();
