+++ src/main/java/org/mo/jims/coop/service/StockReturnService.java
@@ -9,15 +9,13 @@ import org.mo.jims.coop.entity.StockReturn;
 import org.mo.jims.coop.repository.StockReturnRepository;
 import org.mo.open.common.util.Page;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Propagation;
-import org.springframework.transaction.annotation.Transactional;
 
 @Service("stockReturnService")
 public class StockReturnService {
 
 	private StockReturnRepository stockReturnRepository;
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public Page<StockReturn> getStockReturnByCriteria(String goodName,
 			String customerName, Date beginTime, Date endTime, int page,
 			int pageSize) {
@@ -36,12 +34,15 @@ public class StockReturnService {
 		return StockReturnPage;
 	}
 
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public StockReturn getStockReturnByPK(String id) {
-		return stockReturnRepository.selectByPK(id);
+		if(id!=null && !"".equals(id)){
+			return stockReturnRepository.selectByPK(id);
+		}
+		return null;
 	}
 
-	@Transactional(rollbackFor = RuntimeException.class)
+	
 	public boolean saveStockReturn(StockReturn entity) {
 		if (entity != null) {
 			if (entity.getTime() == null) {
@@ -53,109 +54,115 @@ public class StockReturnService {
 		return false;
 	}
 
-	@Transactional(rollbackFor = RuntimeException.class)
+	
 	public boolean alterStockReturn(StockReturn entity) {
-		stockReturnRepository.updateByPK(entity);
-		return true;
+		if(entity!=null){
+			stockReturnRepository.updateByPK(entity);
+			return true;
+		}
+		return false;
 	}
 
-	@Transactional(rollbackFor = RuntimeException.class)
+	
 	public boolean removeStockReturnByPK(String id) {
-		stockReturnRepository.deleteByPK(id);
-		return true;
+		if(id!=null && !"".equals(id)){
+			stockReturnRepository.deleteByPK(id);
+			return true;
+		}
+		return false;
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public List<StockReturn> getStockReturnInToday() {
 		return stockReturnRepository.selectToday();
 	}
 
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public List<StockReturn> getStockReturnInYesterday() {
 		return stockReturnRepository.selectYesterday();
 	}
 
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public List<StockReturn> getStockReturnInThisWeek() {
 		return stockReturnRepository.selectThisWeek();
 	}
 
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public List<StockReturn> getStockReturnInLastWeek() {
 		return stockReturnRepository.selectLastWeek();
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public List<StockReturn> getStockReturnInThisMonth() {
 		return stockReturnRepository.selectThisMonth();
 	}
 
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public List<StockReturn> getStockReturnInLastMonth() {
 		return stockReturnRepository.selectLastMonth();
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public List<StockReturn> getStockReturnInThisYear() {
 		return stockReturnRepository.selectThisYear();
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public List<StockReturn> getStockReturnInLastYear() {
 		return stockReturnRepository.selectLastYear();
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public List<StockReturn> getAllStockReturn(){
 		return stockReturnRepository.selectAll();
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public int countStockReturnInToday(){
 		return stockReturnRepository.countToday();
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public int countStockReturnInYesterday(){
 		return stockReturnRepository.countYesterday();
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public int countStockReturnInThisWeek() {
 		return stockReturnRepository.countThisWeek();
 	}
 
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public int countStockReturnInLastWeek() {
 		return stockReturnRepository.countLastWeek();
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public int countStockReturnInThisMonth() {
 		return stockReturnRepository.countThisMonth();
 	}
 
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public int countStockReturnInLastMonth() {
 		return stockReturnRepository.countLastMonth();
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public int countStockReturnInThisYear() {
 		return stockReturnRepository.countThisYear();
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public int countStockReturnInLastYear() {
 		return stockReturnRepository.countLastYear();
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public int countAllStockReturn(){
 		return stockReturnRepository.countAll();
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public Float getStockReturnPercentageofDayOnThisMonth(){
 		int countToday = stockReturnRepository.countToday();
 		int countYesterday = stockReturnRepository.countYesterday();
@@ -164,7 +171,7 @@ public class StockReturnService {
 		return result;
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public Float getStockReturnPercentageofWeekOnThisMonth(){
 		int countToday = stockReturnRepository.countThisWeek();
 		int countYesterday = stockReturnRepository.countLastWeek();
@@ -173,7 +180,7 @@ public class StockReturnService {
 		return result;
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public Float getStockReturnPercentageofMonthOnThisYear(){
 		int countToday = stockReturnRepository.countThisMonth();
 		int countYesterday = stockReturnRepository.countLastMonth();
@@ -182,7 +189,7 @@ public class StockReturnService {
 		return result;
 	}
 	
-	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	
 	public Float getStockReturnPercentageofYearOnAll(){
 		int countToday = stockReturnRepository.countThisYear();
 		int countYesterday = stockReturnRepository.countLastYear();
